+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 **Features and Improvements**
 
 * Card 193: Add specific fields on several models
+* Card 195: Add translatable flag on several fields
 
 **Bugfixes**